Refactor app boot to prevent translation initialization errors (#1060)
